add team api, changebooking, add team schema, changes in admin routes

+++ database/migrations/20200505165434_guest.js
@@ -9,7 +9,7 @@ exports.up = (knex) =>
     table.string('fullname');
     table.string('country');
     table.string('email');
-    table.integer('phone');
+    table.bigInteger('phone');
     table.date('dob');
     table.string('gender');
     table.string('typeOfDoc');
